test: 불필요한 테스트 제거 및 올바른 위치로 테스트 이동
