+++ src/test/resources/member.sql
@@ -1,0 +1,4 @@
+INSERT INTO member(name, email, password, role)
+VALUES ('운영진', 'admin@naver.com', '1234', 'ADMIN');
+INSERT INTO member(name, email, password, role)
+VALUES ('홍길동', 'member@naver.com', '1234', 'MEMBER');
