+++ src/test/resources/reset.sql
@@ -1,0 +1,9 @@
+DELETE FROM reservation;
+DELETE FROM member;
+DELETE FROM theme;
+DELETE FROM reservation_time;
+
+ALTER TABLE member ALTER COLUMN id RESTART WITH 1;
+ALTER TABLE reservation ALTER COLUMN id RESTART WITH 1;
+ALTER TABLE reservation_time ALTER COLUMN id RESTART WITH 1;
+ALTER TABLE theme ALTER COLUMN id RESTART WITH 1;
