+++ src/test/resources/time.sql
@@ -1,0 +1,4 @@
+INSERT INTO reservation_time(start_at)
+VALUES ('10:00');
+INSERT INTO reservation_time(start_at)
+VALUES ('11:00');
